requirements.txt
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
distlib==0.3.8
filelock==3.15.4
identify==2.6.0
idna==3.10
nodeenv==1.9.1
platformdirs==4.2.2
pre-commit==3.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
urllib3==2.3.0
virtualenv==20.26.3
